+++ inst/shiny_iglu/ui.R
@@ -44,8 +44,10 @@ shinyUI(fluidPage(
                                                                                `MODD` = 'modd',
                                                                                `Quantiles` = 'quantile_glu',
                                                                                `Range` = 'range_glu',
+                                                                               `Rate of Change (ROC)` = 'roc',
                                                                                `Standard Deviation` = 'sd_glu',
                                                                                `Standard Deviation Subtypes` = 'sd_measures',
+                                                                               `Standard Deviation ROC` = 'sd_roc',
                                                                                `Summary Statistics` = 'summary_glu'
                )),
                uiOutput("select_parameter"),
@@ -59,11 +61,14 @@ shinyUI(fluidPage(
                  radioButtons("plottype",  "Plot Type",
                               choices = c(`Time Series` = 'tsplot',
                                           `Lasagna Plot (Multiple Subject)` = 'lasagnamulti',
-                                          `Lasagna Plot (Single Subject)` = 'lasagnasingle'
+                                          `Lasagna Plot (Single Subject)` = 'lasagnasingle',
+                                          `Rate of Change (Time Series)` = 'plot_roc',
+                                          `Rate of Change (Histogram)` = 'hist_roc'
                )),
                uiOutput("plot_lasagnatype"),
                uiOutput("plot_subjects"),
                uiOutput("plot_subjects_help_text"),
+               uiOutput("plot_timelag"),
                uiOutput("plot_maxd"),
                uiOutput("plot_datatype"),
                uiOutput("plot_datatype_help_text"),