+++ NEWS.md
@@ -1,5 +1,11 @@
 # iglu 1.3.0
 
+* Added active percent metric
+* Added cv_measures metrics
+* Added AUC metric
+
+# iglu 1.3.0
+
 * Added glucose management indicator (GMI) and eA1c
 
 # iglu 1.2.0